Delete id not found 404

+++ client_registration/src/main/java/com/oliveiralia/client_registration/service/ClientService.java
@@ -1,14 +1,17 @@
 package com.oliveiralia.client_registration.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.oliveiralia.client_registration.dto.ClientDTO;
 import com.oliveiralia.client_registration.entities.Client;
 import com.oliveiralia.client_registration.repositories.ClientRepository;
+import com.oliveiralia.client_registration.service.exceptions.DatabaseException;
 import com.oliveiralia.client_registration.service.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -41,9 +44,16 @@ public class ClientService {
 		return new ClientDTO(entity);		
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Id not found.");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Id not found.");
+		}		
 	}
 	
 	@Transactional
